rename font-metrics-context.ts to font-data-context.ts
