+++ packages/react/src/index.ts
@@ -4,4 +4,4 @@ import MathmlRenderer from "./mathml-renderer";
 import MathEditor from "./math-editor";
 
 export {MathKeypad, MathRenderer, MathmlRenderer, MathEditor};
-export {FontDataContext} from "./font-metrics-context";
+export {FontDataContext} from "./font-data-context";
